Extract Signal/AbstractSignal to upper level
